Code refactor to support enable/disable indexer or blockchain data service
