+++ services/blockchain-indexer/methods/dataService/controllers/peers.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const dataService = require('../../shared/dataService');
+const dataService = require('../../../shared/dataService');
 
 const getPeers = async params => {
 	const response = await dataService.getPeers(params);
